fix(page props) move PageProps to types folder as gatsby picks it up as a page component
